dataset_analyzer/requirements.txt
pyyaml>=6.0
click>=8.0
openai>=1.0
requests>=2.28
pandas>=1.5
numpy>=1.24
scikit-learn>=1.3
nltk>=3.8
tqdm>=4.64
pathlib2>=2.3
dataclasses-json>=0.5
typing-extensions>=4.0